tests/wpt/tests/css/css-ui/parsing/caret-color-valid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS UI Level 4: parsing caret-color with valid values</title>
<link rel="author" title="Eric Willigers" href="mailto:ericwilligers@chromium.org">
<link rel="help" href="https://drafts.csswg.org/css-ui/#caret-color">
<meta name="assert" content="caret-color supports the grammar '[ auto | <color> ] <color>?'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
// Single-value form.
test_valid_value("caret-color", "auto");
test_valid_value("caret-color", "rgba(10, 20, 30, 0.4)");
test_valid_value("caret-color", "transparent");

// CSS-wide keywords.
test_valid_value("caret-color", "initial");
test_valid_value("caret-color", "inherit");
test_valid_value("caret-color", "unset");
test_valid_value("caret-color", "revert");
test_valid_value("caret-color", "revert-layer");

// Two-value form: the optional second <color> sets the text color overlapping with a
// block caret (https://drafts.csswg.org/css-ui/#caret-color).
test_valid_value("caret-color", "auto auto");
test_valid_value("caret-color", "auto rgba(10, 20, 30, 0.4)");
test_valid_value("caret-color", "rgba(10, 20, 30, 0.4) auto");
test_valid_value("caret-color", "currentcolor currentcolor");
test_valid_value("caret-color", "transparent transparent");
test_valid_value("caret-color", "currentcolor auto");
test_valid_value("caret-color", "auto transparent");
</script>
</body>
</html>
